requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2023.10
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.7
django-cors-headers==4.6.0
django-crispy-forms==2.1
django-environ==0.11.2
django-ratelimit==4.1.0
exceptiongroup==1.2.2
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
openai==1.56.2
packaging==24.2
Pillow==10.4.0
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
stripe==11.3.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
